git-svn-id: svn+ssh://svn.cprover.org/srv/svn/cbmc/branches/peter-incremental-unwinding@3664 6afb6bc1-c8e4-404c-8f48-9ae832c5b171

+++ src/ansi-c/c_typecheck_base.h
@@ -84,14 +84,14 @@ protected:
   {
   protected:
     const exprt array;
-    unsigned pos;
+    size_t pos;
 
   public:    
     explicit init_statet(const exprt &_array):array(_array), pos(0)
     {
     }
   
-    unsigned remaining() const
+    size_t remaining() const
     {
       return array.operands().size()-pos;
     }
