+++ src/solvers/sat/cnf.h
@@ -30,9 +30,9 @@ public:
   virtual literalt limplies(literalt a, literalt b);
   virtual literalt lselect(literalt a, literalt b, literalt c); // a?b:c
   virtual literalt new_variable();
-  virtual unsigned no_variables() const { return _no_variables; }
-  virtual void set_no_variables(unsigned no) { _no_variables=no; }
-  virtual unsigned no_clauses() const=0;
+  virtual size_t no_variables() const { return _no_variables; }
+  virtual void set_no_variables(size_t no) { _no_variables=no; }
+  virtual size_t no_clauses() const=0;
 
   void gate_and(literalt a, literalt b, literalt o);
   void gate_or(literalt a, literalt b, literalt o);
@@ -45,7 +45,7 @@ public:
   static void eliminate_duplicates(const bvt &bv, bvt &dest);
 
 protected:
-  unsigned _no_variables;
+  size_t _no_variables;
   
   bool process_clause(const bvt &bv, bvt &dest);
 
@@ -64,7 +64,7 @@ public:
   {
   }
   
-  virtual unsigned no_clauses() const
+  virtual size_t no_clauses() const
   {
     return clause_counter;
   }
@@ -72,7 +72,7 @@ public:
 protected:
   typedef enum { INIT, SAT, UNSAT, ERROR } statust;
   statust status;
-  unsigned clause_counter;
+  size_t clause_counter;
 }; 
 
 #endif
